fix module counts
